+++ src/dropdown.test.accessibility.ts
@@ -2,7 +2,7 @@ import { expect, test } from '@playwright/test';
 import type Dropdown from './dropdown.js';
 import type DropdownOption from './dropdown.option.js';
 
-test('disabled=${true}', async ({ page }) => {
+test('disabled=${true}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -20,7 +20,7 @@ test('disabled=${true}', async ({ page }) => {
   `);
 });
 
-test('disabled=${false}', async ({ page }) => {
+test('disabled=${false}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
@@ -30,7 +30,7 @@ test('disabled=${false}', async ({ page }) => {
 });
 
 test.describe('filter("add")', () => {
-  test('open=${true}', async ({ page }) => {
+  test('open=${true}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -51,7 +51,7 @@ test.describe('filter("add")', () => {
     `);
   });
 
-  test('open=${false}', async ({ page }) => {
+  test('open=${false}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -68,7 +68,7 @@ test.describe('filter("add")', () => {
 });
 
 test.describe('filter("noMatchingOptions")', () => {
-  test('open=${true}', async ({ page }) => {
+  test('open=${true}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -87,7 +87,7 @@ test.describe('filter("noMatchingOptions")', () => {
     `);
   });
 
-  test('open=${false}', async ({ page }) => {
+  test('open=${false}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -104,7 +104,7 @@ test.describe('filter("noMatchingOptions")', () => {
 });
 
 test.describe('filter("o")', () => {
-  test('open=${true}', async ({ page }) => {
+  test('open=${true}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -127,7 +127,7 @@ test.describe('filter("o")', () => {
     `);
   });
 
-  test('open=${false}', async ({ page }) => {
+  test('open=${false}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -143,7 +143,7 @@ test.describe('filter("o")', () => {
   });
 });
 
-test('hide-label', async ({ page }) => {
+test('hide-label', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -158,7 +158,7 @@ test('hide-label', async ({ page }) => {
   `);
 });
 
-test('loading', async ({ page }) => {
+test('loading', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -175,7 +175,7 @@ test('loading', async ({ page }) => {
 });
 
 test.describe('multiple=${true}', () => {
-  test('open=${true}', async ({ page }) => {
+  test('open=${true}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -206,7 +206,7 @@ test.describe('multiple=${true}', () => {
     `);
   });
 
-  test('open=${false}', async ({ page }) => {
+  test('open=${false}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -229,100 +229,112 @@ test.describe('multiple=${true}', () => {
     `);
   });
 
-  test('<glide-core-dropdown-option>.count', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.count',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.multiple = true;
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.multiple = true;
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.count = 1000;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.count = 1000;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - option /One \\d+\\+/:
-          - checkbox "One"
-        - option "Two":
-          - checkbox "Two"
-        - option "Three":
-          - checkbox "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - option /One \\d+\\+/:
+            - checkbox "One"
+          - option "Two":
+            - checkbox "Two"
+          - option "Three":
+            - checkbox "Three"
+      `);
+    },
+  );
 
-  test('<glide-core-dropdown-option>.disabled', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.disabled',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.multiple = true;
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.multiple = true;
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.disabled = true;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.disabled = true;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - option "One" [disabled]:
-          - checkbox "One" [disabled]
-        - option "Two":
-          - checkbox "Two"
-        - option "Three":
-          - checkbox "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - option "One" [disabled]:
+            - checkbox "One" [disabled]
+          - option "Two":
+            - checkbox "Two"
+          - option "Three":
+            - checkbox "Three"
+      `);
+    },
+  );
 
-  test('<glide-core-dropdown-option>.editable', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.editable',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.multiple = true;
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.multiple = true;
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.editable = true;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.editable = true;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - 'option "One Edit option: One"':
-          - checkbox "One"
-          - 'button "Edit option: One"'
-        - option "Two":
-          - checkbox "Two"
-        - option "Three":
-          - checkbox "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - 'option "One Edit option: One"':
+            - checkbox "One"
+            - 'button "Edit option: One"'
+          - option "Two":
+            - checkbox "Two"
+          - option "Three":
+            - checkbox "Three"
+      `);
+    },
+  );
 });
 
 test.describe('multiple=${false}', () => {
-  test('open=${true}', async ({ page }) => {
+  test('open=${true}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -342,7 +354,7 @@ test.describe('multiple=${false}', () => {
     `);
   });
 
-  test('open=${false}', async ({ page }) => {
+  test('open=${false}', { tag: '@accessibility' }, async ({ page }) => {
     await page.goto('?id=dropdown--dropdown');
 
     await page
@@ -357,87 +369,99 @@ test.describe('multiple=${false}', () => {
     `);
   });
 
-  test('<glide-core-dropdown-option>.count', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.count',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.count = 1000;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.count = 1000;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - option /One \\d+\\+/
-        - option "Two"
-        - option "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - option /One \\d+\\+/
+          - option "Two"
+          - option "Three"
+      `);
+    },
+  );
 
-  test('<glide-core-dropdown-option>.disabled', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.disabled',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.disabled = true;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.disabled = true;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - option "One" [disabled]
-        - option "Two"
-        - option "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - option "One" [disabled]
+          - option "Two"
+          - option "Three"
+      `);
+    },
+  );
 
-  test('<glide-core-dropdown-option>.editable', async ({ page }) => {
-    await page.goto('?id=dropdown--dropdown');
+  test(
+    '<glide-core-dropdown-option>.editable',
+    { tag: '@accessibility' },
+    async ({ page }) => {
+      await page.goto('?id=dropdown--dropdown');
 
-    await page
-      .locator('glide-core-dropdown')
-      .evaluate<void, Dropdown>((element) => {
-        element.open = true;
-      });
+      await page
+        .locator('glide-core-dropdown')
+        .evaluate<void, Dropdown>((element) => {
+          element.open = true;
+        });
 
-    await page
-      .locator('glide-core-dropdown-option')
-      .first()
-      .evaluate<void, DropdownOption>((element) => {
-        element.editable = true;
-      });
+      await page
+        .locator('glide-core-dropdown-option')
+        .first()
+        .evaluate<void, DropdownOption>((element) => {
+          element.editable = true;
+        });
 
-    await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
-      - text: Label
-      - button "Label" [expanded]
-      - listbox:
-        - 'option "One Edit option: One"':
-          - 'button "Edit option: One"'
-        - option "Two"
-        - option "Three"
-    `);
-  });
+      await expect(page.locator('glide-core-dropdown')).toMatchAriaSnapshot(`
+        - text: Label
+        - button "Label" [expanded]
+        - listbox:
+          - 'option "One Edit option: One"':
+            - 'button "Edit option: One"'
+          - option "Two"
+          - option "Three"
+      `);
+    },
+  );
 });
 
-test('placeholder', async ({ page }) => {
+test('placeholder', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -452,7 +476,7 @@ test('placeholder', async ({ page }) => {
   `);
 });
 
-test('select-all', async ({ page }) => {
+test('select-all', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -478,7 +502,7 @@ test('select-all', async ({ page }) => {
   `);
 });
 
-test('slot="description"', async ({ page }) => {
+test('slot="description"', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
@@ -499,7 +523,7 @@ test('slot="description"', async ({ page }) => {
   `);
 });
 
-test('tooltip', async ({ page }) => {
+test('tooltip', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=dropdown--dropdown');
 
   await page
