+++ src/icon-button.test.accessibility.ts
@@ -1,7 +1,7 @@
 import { expect, test } from '@playwright/test';
 import type IconButton from './icon-button.js';
 
-test('disabled=${true}', async ({ page }) => {
+test('disabled=${true}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=icon-button--icon-button');
 
   await page
@@ -15,7 +15,7 @@ test('disabled=${true}', async ({ page }) => {
   `);
 });
 
-test('disabled=${false}', async ({ page }) => {
+test('disabled=${false}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=icon-button--icon-button');
 
   await expect(page.locator('glide-core-icon-button')).toMatchAriaSnapshot(`
