+++ src/tag.test.accessibility.ts
@@ -1,7 +1,7 @@
 import { expect, test } from '@playwright/test';
 import type Tag from './tag.js';
 
-test('disabled=${true}', async ({ page }) => {
+test('disabled=${true}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=tag--tag');
 
   await page.locator('glide-core-tag').evaluate<void, Tag>((element) => {
@@ -15,7 +15,7 @@ test('disabled=${true}', async ({ page }) => {
   `);
 });
 
-test('disabled=${false}', async ({ page }) => {
+test('disabled=${false}', { tag: '@accessibility' }, async ({ page }) => {
   await page.goto('?id=tag--tag');
 
   await page.locator('glide-core-tag').evaluate<void, Tag>((element) => {
